vars/compiletestRunner: Fix path to properties file
The path to the architecture specific properties file does not match
with the effective location of this properties file. A "file does not
exist" warning is the consequence and the architecture specific
compile environment settings are not added to the environment.

Set the path to the effective location to ensure all required
environment settings are done.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>