vars/compiletest*: Cleanup working directory definition
All stages and jobs should be executed in a per job/stage dedicated
workspace directory. The advantage of this clean directory setup is,
that after a whole test run, the workspaces of the jobs and stages
still exist on the nodes and can be used for debugging purpose. The
directory has to be defined explicitly.

Add a dedicated workspace directory definition for the compiletest as
well as for the compiletestRunner.

The compiletestRunner executes the effective kernel compile tasks in a
subdirectory of its workspace. In this subdirectory hierarchy, the
architecture, configuration name and overlay of the kernel that is
built is encoded. It could look like the following:
"results/x86/qemu_x86_64/RT-full/compile". The structure of this
subdirectory hierarchy eases the handling of the test results
artifacts.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>