vars/boottest*: Cleanup workspace directory definition
All stages and jobs should be executed in a per job/stage dedicated
workspace directory. The advantage of this clean directory setup is,
that after a whole test run, the workspaces of the jobs and stages
still exist on the nodes and can be used for debugging purpose. The
directory has to be defined explicitly.

boottest uses a separate workspace directory. boottestRunner can be
executed several times for the same kernel. Therefore adding the
target name to the workspace directory.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>