vars/compiletest: Make proper decision, if boottest will be executed
The decision, if a kernel will be boottested after the build, depends
on the existance of a *.properties file. If this file is available, a
boottest needs to be executed after the compile test. The current
implementation executes the boottest unconditional and hands in an
empty string if no boottest needs to be executed. Furthermore, the
properties file is added unconditionally to the environment which
could lead to the warning "file does not exit".

Make a proper decision, based on the existance of the properties file,
if the compile and boottest or only compiletest is exectued.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>