vars/compiletest: Move environment checking to a single place
Environment checking is done in two functions (in call() and in
compile()).

Move the dereferencing environment parameters directly into the try
area and pass through the results to compile() method as argument
eases the readability. Explain the reason for the nullpointer try
catch statement in call() method in a comment.

A better solution would be to check the environment in a separate job,
but that's outside the scope of this change. Add a comment to that
effect and mark it TODO.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>