vars/cyclictest*: Clean workspace directory handling
All stages and jobs should be executed in a per job/stage dedicated
workspace directory. The advantage of this clean directory setup is,
that after a whole test run, the workspaces of the jobs and stages
still exist on the nodes and can be used for debugging purpose. The
directory has to be defined explicitly.

cyclictest uses a separate workspace directory. cyclictestRunner can
be executed several times on the same target. Therefore adding the
cyclictest name to the workspace directory.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>