vars/targetprep: Use dedicated workspace directory
All stages and jobs should be executed in a per job/stage dedicated
workspace directory. The advantage of this clean directory setup is,
that after a whole test run, the workspaces of the jobs and stages
still exist on the nodes and can be used for debugging purpose. The
directory has to be defined explicitly.

Add a dedicated workspace directory for targetprep.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>