vars/boottestRunner: Inline reboottarget script
This script is part of the external resources. The outsourcing of
functionality in scripts makes the code harder to read and harder to
debug.

Therefore inline the script. Write the script into a file for
debugging purpose. The rm of the files is not required, because the
workspace is cleaned up before executing the task.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>