vars/boottestRunner: Splitting boot log at new delimiter in groovy
The serial console log was cutted at the last bootpart, which starts
at [ 0.000000]. This is not working, when reboot fails and target
boots into the default kernel and the serial console log is recorded
further.

Therefore, a echo to "/dev/kmsg" is done on the target, when it
executes a kexec into the test kernel. The splitting of the bootlog is
now implemented in groovy with the help of the new delimiter. The part
of the bash script is removed.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>