vars/compiletestRunner: Integrate compile script in groovy
The compile script, that creates the compile command script, is stored
as an external resource bash script. The external resource bash script
is executed by using the helper class. The helper class injects
several environment settings before executing the script. This makes
the code harder to read and harder to debug.

Integrate the creation of the compile bash script in groovy. Add the
required environment variables to the script.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>