vars/compiletestRunner: Implement creation of compile command in groovy
The bash script compile_script.sh, that is exectued, creates another
bash script, which holds the effective compile command. This makes the
code hard to read.

Create only the script with the compile commands as a bash script,
that is afterwards exectued using the sh() command. Implement the
required logic to create this script in groovy.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>