vars/compiletestRunner: Use dedicated build and result directory
The name of the result and build directory are stored as environment
in the helper environment class. When a script is exectued by using
the helper class, the environment is injected before the script is
exectued. This mechanism impairs the debugging and readability of the
code so the aim is to get rid of this mechanism.

As a preparation use groovy variable for the dedicated compile and
result directory name. While at it, rename the bash script variable
with more self explaining names.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>