vars/compiletestRunner: Use a variable for make parameters
Make is called in the compile script several times with the same
parameters. Ease the readability and improve robustness of the script
and use a single variable which combines all required
parameters. While at it convert all different parameters to
LOCALVERSION, which match in the end on the Jenkins environment
variable BUILD_NUMBER.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>