vars/compiletestRunner: Use a variable for logfile
Content is added to the logfile at different locations in the
script. Use a variable instead hardcoding the name of the logfile
every time. To handle all logfiles in the same way, use a variable for
the package build logfile as well and add a comment to the first line
of the package log file.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>