vars/compiletestRunner: Ease CONFIGNAME handling
The combination of configuration and overlay is required in several
places of the compile script. This information is generated with
different combinations of architecture, configuration name and
overlay. Adapt the notation to be consistent within the script.

Add a TODO which is related to this notation, that depends on the
restructuration of the test environment building jobs.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>