vars/compiletestRunner: Add additional test and TODOs to compile script
When the compile script is used to reproduce the compile test, the
same directory structure is required. To ensure that the build and
result directories exist, add a test. Add also a reminder that the
proper configuration needs to be stored at a defined place and the
patches need to be applied.

Furthermore add TODOs how the compile script could be improved in the
future.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>