vars/compiletestRunner: Execute prepare kernel config script with parameters
This script is part of the external resources and is executed via the
helper class. The helper class therefore injects several environment
variables before executing the script. This makes debugging problem
with environment settings harder to debug.

Therefore execute the prepare kernel config script with parameters and
call it directly using the sh() command in groovy.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>