vars/cyclictestRunner: Implement cyclictest.sh in groovy
cyclictest.sh is stored as a resouce and is executed with helper
class, so that environment is setup properly. It creates the script
histogram.sh and executes it afterwards. This outsourcing of
functionality makes the code harder to read and harder to debug.

Inline creation of histogram.sh and execute it with 'sh' command.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>