vars/CIRTinputcheck: Add a dedicated groovy class for inputcheck
Inputcheck examines, if all parameters are set properly. It is done
via a groovy source file. If a Jenkinsfile should directly call
inputcheck, outsourcing it into a dedicated class eases the
call. Otherwise the path needs to be imported.

The test itself defines on which node it has to run. Therefore add the
node{} closure. Furthermore GUI_TESTDESCR_BRANCH is added to the
handed in map, so only a single call() method is required.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>