vars/boottestRunner: Split out reboot target
The functionallity to reboot a target is implemented in shell. The
shell scripts should be removed. Therefore implement the reboot target
functionality into an own groovy function. Move also the offline()
call, which marks the target as offline, into this new function.

While at it, add several important comments to the code.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>