vars/boottestRunner: Add error handling for virsh consolelog call
virsh consolelog is executed in background. If there occurs an error
during logging of the console, the error is detected only when the
command will be killed.

An alternative command to the virsh consolelog command is added and
the stderr of virsh call is piped into an error file. Wait some
seconds before the error file is examined, to be sure that the error
already occurs or everything was fine.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>