vars/boottestRunner: Move archive artifacts in a finally section
Archive artifacts is not executed, when an error occurs in
runner().

Move the archiveArtifacts() call into the finally section of the
try-catch statement around the runner call. The archiveArtifacts has
to know the location of the artifacts. The environement handling is
moved into the call() function and all required parameters are handed
into the runner().

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>