vars/boottestRunner: Reboot into default kernel in finally
Reboot into default kernel has to be done in any case. Therefore move
it into the finally section and add a try catch statement arround the
important calls. The reboot has to be done as long as the test is
holding the target lock.

To reuse existing functions rebootTarget() and checkOnline() a boolean
parameter is added, that specifies if it is a test boot or not.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>