vars/boottestRunner: Execute cyclictest conditional
Cyclictest test is called, even if there is no cyclictest defined.

Cyclictest call needs to be conditional. While at it remove the
directory comment and the node{} closure, because the node the test
has to be executed on is defined inside the test itself. The comment
is then no longer required.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>