vars/CIRTbuildenv: Add a dedicated function to read "clean" files
Test description files are transformed in *.properties files, which
contain the proper environment setting for a specific test (for
example compiletest of x86 defconfig Kernel configuration). The
*.properties files should be easily parsable. The empty lines and
comment lines are not required. Comment lines start with a #.

Therefore create a readCleanFile() function which removes the empty
and comment lines and call this function instead of readFile().

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>