vars/CIRTbuildenv: Allow test-description without patches directory
The test-description does not necessarily contain a patches
directory. In this case the stash patches call will fail. If only a
query would be added, to test if the patches directory exists, the
unstash call of patches stash will fail as well because of not
existing stash.

Therefore add the parameter "allowEmpty: true" to create an empty
stash without a patches directory.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>