Add shunit command
To perform tests and create JUnit result files pyjutest is used. pyjutest
creates a junit xml based on stdout and stderr of a passed command.

This implements a wrapper function that ensures that a given command is
passed to run with pyjutest wrapped.

ensureDebPkg is used to ensure that pyjutest is installed on the targetnode.
ensureDebPkg is not able to support different suites on the targetnodes
atm.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Manuel Traut <manut@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>