vars/compiletestRunner: Stash also *.tar.xz for devicetree binaries
Devicetree binaries are required for the targetprep of the
boottest. targetprep unstashes the 'linuximage' stash and is looking
for the archive with devicetree binaries and unpack it. If the device
tree is not available but required, boottest will fail.

While at it, use the variable for the resultdir instead of hardcoding
it.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>