vars/cyclictest: Integrate cyclictest2xml.py to generate junit xml files
Call the python script cyclictest2xml.py after a run of cyclictest to
parse the output. The resulting xml file is then put into a stash.

It is executed in cyclictest instead of cyclictestRunner, so that the
cyclictest2xml.py script is executed on master on not on the test
target.

[anna-maria: use cyclictest2xml.py instead of collect.py and adapt
commit message]

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>