vars/boottestRunner: Create xml from boottest, collect files for database
To capture the result of the boottest, this step uses junit xml as
output format like the other tests. This is accomplished by calling
a python script (resources/.../boottest2xml.py) which puts the bootlog
into an xml file. The script is called after it is determined if
the boot was successful or not.

Write the cmdline of the test device only to the file "cmdline", if
the device is in testing mode (if testboot is set), so that the file
doesn't get overwritten when the default kernel is booted again.

[anna-maria: do not store default kernel cmdline]

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>