resources/.../feedDatabase.py: Write compiletest result to database
As this script gets called once per compile configuration, read the
result from the xml file and the additional information from files and
environment and write it to the database.
Collect the success state of all tests and update the "pass" flag in
the cirtscheduler table accordingly. If the script gets called for the
last time, which is signaled via a command line parameter, also update
the "processing" flag to False.

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>