vars/feeddatabase: Add logic to call feedDatabase python script
This script collects all environment data, relevant build artifacts
and junit xml files from the tests and then calls the feedDatabase
python script for every compiletest configuration.
As the python script needs to setup the cirtbranch and cirtscheduler
tables only once and in the end set "processing" to false, it is
signaled via comand line arguments whether it is the first or the last
run (if there was only one compiletest run, these two cases can happen
at the same time).

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>