resources/.../feedDatabase.py: read cyclictest script directly
The current implementation writes the cyclictest script as attribute
into the junit xml file. Either in the read or the write step the line
breaks seem to disappear.
As the script is stashed anyway, we can read it in the database
step directly for now, until we find a better way.

[anna-maria: add TODO comment and fix proper encode]

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>