boottestRunner: Do not use currentBuild.result
To be sure to parse only the state of the current boot job "failed"
needs to be parsed. "currentBuild.result" cannot be used, since it is
a global state and is influenced by the state of concurrent
builds. The call of junit() is not removed, because it sets the
Jenkins internal overall state in a proper way.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>