junit_return: Introduce junit call with return values
junit() calls mark the currentBuild.result as UNSTABLE if the junit
file has a "failed" tag. The currentBuild.result is a global variable
and the result of a single stage cannot be represented. Therefore
create a new global variable that examines the occurence of a
"failure" in the junit xml file. Set a return state to 'UNSTABLE' or
'SUCCESS' in dependecy of this occurence.

Make shunit to use junit_result() instead of junit().

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>