feeddatabase: Ignore non existing stash due to failure
In case of a compile or boot failure one or more stashes with test
result data may not exist. Catch the abort exception thrown by unstash
in case of a non existing stash, report about it and go further.

While at it, adapt error handling of feeddatabase in the same way it
is done at other test with a try-catch statement.

[anna-maria: fix whitespace damage and adapt commit message]

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>