Change cirtbranch table
To get a proper warnings history the scheduler job was reintegrated
into CIRT-libs. As an impact the former 1 to 1 relationship of Jenkins
build number and primary key id number in the database has to be
droped. The table "cirtbranch" maps to the Jenkins test description
branch with commit, build number and primary key id number has been
set up.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Sigend-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>