vars/CIRTscheduler: Implement overall email notification
CIRTscheduler should send a overall information about the build state
in anycase to the tester. Therefore the recipients are required. Adapt
CIRTbuildenv to return the recipients to CIRTscheduler so that the
helper is not required in this context.

Furthermore a email is sent to the email addresses which are stored in
GUI_FAILURE_NOTIFICATION in case there was a problem with the test
infrastructure.

[anna-maria: split it into a separate patch; fix typos]

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>