vars/compiletestRunner: fetch git repo history only when required
With cloning depth of 1, git describe works only if the latest commit
is a tag. Fetching the whole history precautionary takes a lot of
time, which is not required.

Add a comment above checkout() function and add a git fetch when git
describe would not work. Fetching with a depth of 1000 should be
enough. The RT patch queue does not have more than 1000 patches and
had less in the past.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>