vars/boottestRunner: Use local variables
Without a type definition or a def before a local variable, the
variable is global an could be accessed and overwritten by another
context. To prevent this behavior use only local variables in
functions.

While at it, tighten the variable scope of some local variables and
solve unattended global variable scope by adding the necessary
parameter to the function signature.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>