resources/.../feedDatabase: Store scheduler_id
Since commit 93a0455cfaa3 ("Change cirtbranch table"), scheduler_id is
autoincrement only and is not the Jenkins build number. Therefore the
scheduler_id needs to be stored in the first run of feedDatabase, to
be able to make references to this autoincrement number.

This patch should be improved, by executing a database query instead.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>