resources/.../feedDatabase.py: Fix not subscriptable object
When feedDatabase is executed more than one time, "pass" row entry of
cirtscheduler table needs to be updated. Therefore the current value
of "pass" needs to be concatenate with the remaining pass values. The
SQLalchemy query returns a class object of type "cirtscheduler"
instead of a dictionary. Fix the access and use the "pass_" member of
the class.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>