shunit: Execute python3 with forced unbuffered I/O stdout stream
When executing compiletest with shunit and pyjutest,
java.io.NotSerializableException of hudson.model.Hudson was raised
from time to time. With additional debug output the following
information about the exception could be generated:

"wrapper script does not seem to be touching the log file in <path to
log file> (JENKINS-48300: if on a laggy filesystem, consider
-Dorg.jenkinsci.plugins.durabletask.BourneShellScript.HEARTBEAT_CHECK_INTERVAL=300)"

The compiletests generate a lot of I/O traffic. stdout over
non-interactive ssh is buffered and the Jenkins logfile is possibly
not feeded with new data in the given check invervall. Jenkins assumes
the jobs are no longer working and aborts their execution.

Force pyjutest to use unbuffered I/O stdout stream to ensure, that the
log file is feeded with new data within the given check interval

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>