vars/notify: Enable some special env variables in email templates.
Jenkins env variable is not a map. Some members of env are given through
callback functions. These special cases are not part of the Jenkins build
job environment and not accessable in email templates.
GIT_COMMIT and GIT_URL are some of these special cases. To be able to
access those variables within a template, add them to the variable
substitution map.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>