CIRT-scheduler: Send notification at begin of test run
Without an initial notification that the test run starts, a user
without access to the Jenkins master instance has no information if
test was successfully triggered. Therefore add a notification email
before compile-, boot- and cyclictests are started. It cannot be done
before parsing the test-description, because then recipients are
unknown. For the implementation a template as well as a variant of the
notify call() function is added.

Signed-off-by: Anna-Maria <anna-maria@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>
Tested-by: Benedikt Spranger <b.spranger@linutronix.de>