compiletestRunner: Exclude debug kernel packages from stash
If CONFIG_DEBUG_INFO is set, a debug kernel package is build. These
packages be prone to be big. They are required for debuging purpose
but there is no need to install them on the target. Add kernel debug
packages to artifacts, but exclude them from the stash transferred to
target.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>