vars/compiletestRunner: Fix boottest dependent debian package build
If there exists a kernel configuration corresponding *.properties
file, the kernel will be boottested after the compile. Only for
boottested kernel configs the debian packages are built, because
package builds take some extra time and need a lot of storage. There
was an error in the query and for every kernel config the debian
packages were built.

While fixing the query, rename the non obvious parameter name and add
a comment.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Manuel Traut <manut@linutronix.de>