notify: Split out non serializable functions
Access to Hudson.instance.getRootUrl() may fail with a
java.io.NotSerializableException. Split call out in a helper function
and mark it @NonCPS. While at it split out access to
StreamingTemplateEngine, which seem to be error prone too. Put the
template preconditioning into the same helper function for better
readability and structure.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>