boottestRunner: Split failure notification
The failure notification helper failnotify() of boottestRunner uses a
hardcoded subject and template. Additional failure notifications in
boottestRunner will need a different subject and template, but the same
preperatory work. Make subject and template configurable in failnotify()
and add an additional signature to pass additional values to notify()
through failnotify().

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>