vars/boottestRunner: Fix RebootException notification
The call of failnotify() used the wrong arguments which leads to a
java.lang.NoSuchMethodError. Remove the wrong "global" argument.

Variables branch, config and overlay are required for RebootException
notification call. Therefore do not unset config and overlay and set
the branch variable via getVar() call.

The build will not fail (and in database the test is successful), but
it is good to show in the overall test result that there was a
problem. Mark the test as UNSTABLE.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>