Enable regression tests for forced reboot
To be able to test the forced reboot exception code paths, the target
is not allowed to boot after a failed soft boot or a failed reboot.

Therefore in test-description the REGRESSION_TEST_FORCED_REBOOT
parameter is set in env/global file, to prevent in forced reboot path
execution of virsh destroy and start.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>