resources/.../preparekernel: Distinguish script return code
preparekernel script may fail for a couple of different reasons. For now
the script failed with exit code 1 for all reasons.

Distinguish the script return code to be able to differ the different
failures from calling side.
While at it use exit instead of return to mark explicit exit points and
do not rely on "set -e".

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>